+++ apps/fluent-tester/src/FluentTester/TestComponents/MenuButton/testData.ts
@@ -3,6 +3,7 @@ import { MenuButtonItemProps } from '@fluentui/react-native';
 import { SvgIconProps } from '@fluentui-react-native/icon';
 export const testImage = require('../Button/icon_24x24.png');
 import TestSvg from '../Button/test.svg';
+import { MENU_ITEM_1_COMPONENT } from './consts';
 
 const svgProps: SvgIconProps = {
   src: TestSvg,
@@ -14,6 +15,7 @@ export const menuItems: MenuButtonItemProps[] = [
     itemKey: '1',
     text: 'MenuItem 1',
     icon: testImage,
+    testID: MENU_ITEM_1_COMPONENT,
   },
   {
     itemKey: '2',