+++ apps/ios/CHANGELOG.md
@@ -1,8 +1,48 @@
 # Change Log - @fluentui-react-native/tester-ios
 
-This log was last generated on Tue, 25 Jan 2022 21:55:31 GMT and should not be manually modified.
+This log was last generated on Wed, 02 Feb 2022 02:29:07 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.1
+
+Wed, 02 Feb 2022 02:29:07 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.52.0
+
+## 0.0.1
+
+Fri, 28 Jan 2022 20:52:39 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.51.5
+
+## 0.0.1
+
+Thu, 27 Jan 2022 10:39:35 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.51.4
+
+## 0.0.1
+
+Wed, 26 Jan 2022 20:31:04 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.51.3
+
+## 0.0.1
+
+Wed, 26 Jan 2022 17:02:27 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/tester to v0.51.2
 
 ## 0.0.1