+++ packages/components/Button/CHANGELOG.md
@@ -1,8 +1,38 @@
 # Change Log - @fluentui-react-native/button
 
-This log was last generated on Tue, 25 Jan 2022 21:55:30 GMT and should not be manually modified.
+This log was last generated on Wed, 02 Feb 2022 02:29:07 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.21.19
+
+Wed, 02 Feb 2022 02:29:07 GMT
+
+### Patches
+
+- Bump @uifabricshared/foundation-compose to v1.11.9
+- Bump @fluentui-react-native/icon to v0.11.6
+- Bump @fluentui-react-native/interactive-hooks to v0.12.2
+- Bump @fluentui-react-native/pressable to v0.8.18
+- Bump @fluentui-react-native/text to v0.11.9
+
+## 0.21.18
+
+Wed, 26 Jan 2022 20:31:04 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.12.1
+- Bump @fluentui-react-native/pressable to v0.8.17
+
+## 0.21.17
+
+Wed, 26 Jan 2022 17:02:27 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.12.0
+- Bump @fluentui-react-native/pressable to v0.8.16
 
 ## 0.21.16